requirements.txt
sqlalchemy==1.4.46
fastapi==0.89.1
alembic==1.9.1
uvicorn==0.20.0
python-dotenv==0.21.0
psycopg2-binary==2.9.5
pytest==7.2.0
asyncpg==0.27.0
pydantic==1.10.4
httpx==0.23.3
aioredis==2.0.1
sqlalchemy_utils==0.39.0
pre-commit==3.0.1
pytest_asyncio==0.20.3
celery==5.2.7
kombu==5.2.4
XlsxWriter==3.0.8
